+++ homeassistant/util/enum.py
